Tanks and APCs
Just did the Tanks and APCs to start going to look at the fliers and the motorcycles later. Fixed my misunderstanding that blunt should be higher than sharp protection.  Removed the Ammo rack in the Taurox as it doesnt have a turret so no need for it to carry ammo.